+++ src/file/PreCacheFile.cc
@@ -35,7 +35,7 @@ void PreCacheFile::run()
 	if (size < 1024 * 1024) {
 		// only pre-cache small files
 
-		const unsigned BLOCK_SIZE = 4096;
+		const size_t BLOCK_SIZE = 4096;
 		unsigned block = 0;
 		unsigned repeat = 0;
 		while (true) {
